tests

+++ package_extension.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
+# Run tests first
+echo "Running tests..."
+npm test
+if [ $? -ne 0 ]; then
+    echo "Tests failed. Aborting package creation."
+    exit 1
+fi
+
 # Remove any existing package
 rm -f extension.zip
 
@@ -13,6 +21,10 @@ cd src && zip -r ../extension.zip \
     js/ \
     -x "*.DS_Store" \
     -x "*/.git/*" \
-    -x "*/.idea/*"
+    -x "*/.idea/*" \
+    -x "*/tests/*" \
+    -x "*/node_modules/*" \
+    -x "*/package.json" \
+    -x "*/package-lock.json"
 
 echo "Extension has been packaged as extension.zip"
